chore(nx-plugin): fix linting errors and standardize structure
With the introduction of Nrwl building Nx with Nx, a lot of linting errors were created. This change fixes all of the schematic and builder linting errors in the nx-plugin package, and also updates the file structure to match more recent changes to the code base.
