+++ packages/nx-plugin/src/schematics/e2e-project/lib/update-nx-json.ts
@@ -1,0 +1,10 @@
+import { Rule } from '@angular-devkit/schematics';
+import { addProjectToNxJsonInTree } from '@nrwl/workspace';
+import { NxPluginE2ESchema } from '../schema';
+
+export function updateNxJson(options: NxPluginE2ESchema): Rule {
+  return addProjectToNxJsonInTree(options.projectName, {
+    tags: [],
+    implicitDependencies: [options.pluginName],
+  });
+}
