+++ packages/nx-plugin/src/schematics/e2e-project/schema.d.ts
@@ -1,9 +1,14 @@
-import { Linter } from '@nrwl/workspace';
-
 export interface Schema {
   pluginName: string;
   npmPackageName: string;
   pluginOutputPath: string;
   jestConfig: string;
   tsSpecConfig: string;
+}
+
+export interface NxPluginE2ESchema extends Schema {
+  projectRoot: string;
+  projectName: string;
+  pluginPropertyName: string;
+  npmScope: string;
 }
