+++ packages/nx-plugin/src/schematics/plugin/lib/update-tsconfig.ts
@@ -1,0 +1,16 @@
+import { Rule, Tree } from '@angular-devkit/schematics';
+import { getProjectConfig, updateJsonInTree } from '@nrwl/workspace';
+import { NormalizedSchema } from '../schema';
+
+export function updateTsConfig(options: NormalizedSchema): Rule {
+  return (host: Tree) => {
+    const projectConfig = getProjectConfig(host, options.name);
+    return updateJsonInTree(
+      `${projectConfig.root}/tsconfig.lib.json`,
+      (json) => {
+        json.compilerOptions.rootDir = '.';
+        return json;
+      }
+    );
+  };
+}
