+++ packages/nx-plugin/src/schematics/schematic/schema.d.ts
@@ -4,3 +4,12 @@ export interface Schema {
   description?: string;
   unitTestRunner: 'jest' | 'none';
 }
+
+export interface NormalizedSchema extends Schema {
+  fileName: string;
+  projectRoot: string;
+  projectSourceRoot: string;
+  npmScope: string;
+  npmPackageName: string;
+  fileTemplate: string;
+}
