+++ packages/nx-plugin/src/utils/get-file-template.ts
@@ -1,0 +1,7 @@
+import { stripIndents } from '@angular-devkit/core/src/utils/literals';
+
+export function getFileTemplate() {
+  return stripIndents`
+      const variable = "<%= projectName %>";
+    `;
+}
